# docs/api-docs/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "中间层模块",
        "description": "虚拟宠物医院API文档，前端用。字段默认值：int为-1，String为'This field is currently Empty.'，Date为null，应该会表现为空字符串。",
        "version": "1.0.1-SNAPSHOT"
    },
    "servers": [
        {
            "url": "http://pikapikpinoimac:8090",
            "description": "Generated server url"
        }
    ],
    "paths": {
        "/api/system/room": {
            "get": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "获取科室信息列表接口",
                "operationId": "getRooms",
                "parameters": [
                    {
                        "name": "roomRole",
                        "in": "query",
                        "description": "科室人员角色，NRAD顺序表示无权限、前台、医助、医师",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "roomList": [
                                            {
                                                "roomRole": "this is a test",
                                                "roomName": "testroom"
                                            }
                                        ]
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "更改科室信息接口",
                "operationId": "updateRoom",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "更改后的科室信息",
                                "example": {
                                    "previousRoomName": "test room",
                                    "roomName": "Prepare",
                                    "roomRole": "RAD"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "roomRole": "RAD",
                                        "roomName": "Prepare"
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "新增科室信息接口",
                "operationId": "addRoom",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "json",
                                "description": "新增的科室信息",
                                "example": {
                                    "roomName": "Prepare",
                                    "roomRole": "RAD"
                                }
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "roomRole": "RAD",
                                        "roomName": "Prepare 2"
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/system/room/{roomName}": {
            "delete": {
                "tags": [
                    "系统管理模块"
                ],
                "summary": "删除科室信息接口",
                "operationId": "deleteRoom",
                "parameters": [
                    {
                        "name": "roomName",
                        "in": "path",
                        "description": "科室信息Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "515": {
                        "description": "Request failed",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 515,
                                    "message": "Request failed"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "example": {
                                    "code": 200,
                                    "data": {
                                        "roomRole": "this is a test",
                                        "roomName": "testroom"
                                    },
                                    "message": "ok"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}